docgen: Expand createSelector() pattern matching to recognize getDependants
In #40025 we ran into a situation where a `createSelector()` selector is
building an opaque value memoized on state values as dependencies. While
this is an unexpected use of a selector it's legitimate and required leaving
some additional code and an explanatory comment in order to avoid breaking
the `docgen` process.

In this patch we're adding recognition for that second argument to the
`createSelector()` function, the `getDependants()` function, and if that
function has more parameters than the `selector` itself we'll cheat
and act like its parameters were listed on the selector. This will
likely only happen in practice when the selector ignores `state` but
it's pluasible someone might go further and use other inputs in the
dependency selection but ignore them on the actual state creation.